refactor(main-navigator): move to screen folder
