+++ src/screens/main-navigator/styles.js
@@ -1,0 +1,14 @@
+import Colors from '../../constants/colors';
+
+const ScreenOptions = {
+  headerStyle: { backgroundColor: Colors.blue },
+  headerTintColor: Colors.white,
+  contentStyle: { backgroundColor: Colors.white },
+  headerTitleAlign: 'center',
+  gestureEnabled: false,
+  headerShown: false,
+};
+
+export default {
+  ScreenOptions
+};
